Major syntax highlight update
This makes some major changes to the syntax highlighting, adding
keywords, correcting some syntax class discrepancies, adding some
handling for labels, builtins, and other pre-declared identifiers, and
removes operators.

Operator removal is following a precedent in vim-go, where this is a
toggle - rather than having that setting for now, I've just removed it.
It can be re-added and put behind a toggle if need be.

This also adds spell checking to only select groups (comments and
strings).