Display fmt errors on save
Currently, "sentinel fmt" on save fails silently, this just adds some
basic messaging when there is an error so you know that there is
actually one. We will want to fix this to some better syntax checking
and general policy verification when we have it.

Note that this currently will display ANSI escape characters in the
output, a further release of Sentinel should keep this from happening by
disabling color when there isn't a TTY.